Fix serial communications when using a serial port that doesn't support Space/Mark parity
Because transmitted bytes go into a buffer, there is no guarantee that setting the parity to EVEN/ODD will affect the correct byte. As a hacky workaround, we insert a small delay (roughly half a byte worth) to give time for the byte to make it out of the buffer before the parity is changed.

Unfortunately I can't see a way to wait for the byte to be transmitted before proceeding to the next byte.